Refactor temporary redraw disablement to use new utility method
Temporary redraw disablements via Control#setRedraw(boolean) have to
ensure that redrawing is enabled in the end. This requires, first, to
always have pairs of calls to setRedraw, and, second, to ensure that
re-enablement is executed even if the code in between throws an
exception.

A utility method added to SWT ensures this and is applied to most
temporary disablements of redraw in this repository.