Ensure consistent coloring of FindReplaceOverlay
When applying custom background colors to editors, it is possible that
the FindReplaceOverlay does not adopt these customized colors in a
consistent way also leading unreadable text.

This change ensure that configured background colors are consistently
applied to the FindReplaceOverlay, i.e., that the background color for
all composites and the text fields is the same.