[GTK] Fix vertical cut-off in input fields of find/replace overlay +2955
The input fields used in the FindReplaceOverlay do currently not exceed
the vertically available size but are just vertically centered in their
default size. Particularly on Linux/GTK this leads to a cut-off, so that
some characters are not visible completely or at all. For example, the
underscore is not visible in some themes.

This change makes the input fields exceed the vertically available size,
such that all input is properly visible.

Fixes https://github.com/eclipse-platform/eclipse.platform.ui/issues/2955