Allow to ignore explicitly defined disabled icons
The algorithm for generating disabled icons on-the-fly has recently been
enhanced. Most explicit disabled icons that have been embedded into
bundles conform to what now can be generated on-the-fly. In addition,
the algorithm is interchangeable, allowing custom stylings of disabled
icons. However, when there are still bundles, such as extensions out of
own control, that still explicitly define disabled icons, exchanging the
algorithm for disabled icons will lead to inconsistent appearance as
only the on-the-fly generated icons will adhere to that.

This change allows to ignore explicitly defined disabled icons, such
that even if some bundles defined disabled icons, they will be ignored
and on-the-fly generated disabled icons according to the selected
algorithm will be used instead.
# Conflicts:
#	bundles/org.eclipse.jface/META-INF/MANIFEST.MF